@shalinilohia50 2025-01-06
> 1. Please add the number and kind of objects used for performance testing.
> 2. Please add the execution time with/without JOIN in the view without the internal function we have introduced.
> 3. Please create a task to look for possibility of bringing down the performance numbers.
> 4. Please also add perf numbers for sp_table_privileges procedure.

2. Not adding the numbers for without JOIN since it would give incorrect results.